[TASK] Add missing CSS and icons for map
It appears that it was missed to add and include the CSS
used for styling the map. Beforehand the CSS was loaded
externally and while restructuring the templates it was
forgotten to add and include them locally again. This is
now resolved and the map should work as intended again.