[TASK] Reduce pipeline usage
It is more than enough to test with lowest and highest
php-version per TYPO3 version and leaving versions in
between out. The hightst php-version reports and fails
on all issues for not tested version in the range in
general.

That reduces the executed pipeline runs which saves some
resources (green it) and maginally speeds up pull-request
pipeline execution times.