[BUGFIX] Use suitable default for flexform select field
TYPO3 v12 flexform definition declared invalid default
value for the `settings.viewMode.default`, which does
not exists in the available items.

This change modifies the flexform definition to declare
`list` instead of invalid `asc` default value for the
`settings.viewMode.default` of the TYPO3 v12 flexform
and matching the TYPO3 v13 flexform definition.