[TASK] Remove output of not-existing property `Profile->slug`
The `EXT:academic_persons` domain model `Profile` does not
have a property `slug` or getter/setter albeit having that
database field and displaying it for profile show view as
part of `EXT:academic-persons-edit` does not make sense,
it will be always empty.

This change removes the obsolete output for now and can be
re-added after technical decision has been made if `slug`
should be part of the domain model or not.