[TASK] Enrich template variables with content elemen data
Sometimes there is the need for rendering informations contained
in the content element which delivers the list plugin. This could
be headers or layout informations just to name two of them. For
ease of use we inject those information into the view variables.

This gives an easy way for rendering those stuff, e.g:
```
<f:render partial="Header/All" arguments="{data}" />
```