[BUGFIX] Ensure correct sorting value persisting new extbase models (FE)
`EXT:academic_persons` provides a couple of domain models defining the
buisness domain modeling (BDD) for the database records including that
relation records needs sorting values to ensure a deterministic order,
otherwise it would lead to a lottery in display.

In BE context this is usually not that of an issue as TYPO3 DataHandler
takes care of sorting fields, special when they are configured correct,
and is uses as part of the backend FormEngine.

`EXT:academic_persons_edit` providies the capability to list and manage
`EXT:academic_persons` data records in the frontend based on the profile
of the logged in user, so-called self-management or profile management.

The default views provides list and the ability to sort related records,
for example additional provile informations, contracts and others next
to default CRUD operations splitted in suited controllers and the same
logic for the `sortAction` has been provided to all of the controllers
for each records.

Not leading to an exception, it has not been discoverd that sorting does
not work as expected until a recently cross check as part of a detailed
developer based test round-up check.

The investigation showed that the issue is composed of two dedicated
issues:

*   TCA needs correct configuration of `TCA->ctlr->sortBy` and must have
    a `TCA->column-><sortByField` with a TCAtype which allows persisting
    and reading the sorting field data, which excludes `TCAtype=none`.
    This has been fixed already throughout the `EXT:academic_persons` in
    previous change.

*   With `TCA->ctrl->sortBy` TYPO3 takes care of correct sorting field
    value handling in the backend (FormEngine and DataHander), which is
    not a thing for the frontend or extbase in general (FE/BE) and that
    means that this needs to be handled manually within executing code,
    like controller actions.

*   Similar to the lack of `sortBy` field handling within extbase for
    resorting, newly created records (models) needs to set a correct
    sorting value manually. If that is not done, all records using
    extbase ORM will create `0` as sorting value for each record and
    breaking sorting handling for the user.

To mitigate the remaining above mentioned issues this change modifies
all `EXT:academic_persons_edit` controller actions persisting the new
models and the sorting actions to accomplish following:

*   Persisting new model actions caluclates the highest `sortingValue`
    for existing relation entities and set it incremented by one (+1).
    Except profiles itself, which are not createable in the frontend
    by the frontend user.

The sorting action needs to be evaluated in a dedicated change if more
optimization are possible, for example detecting and dealing in case
two records have the same sorting value and would lead to a no-op not
obvious for the frontend user.