[BUGFIX] Add missing delete confirmation template
The fluid template for the contract deletion confirmation
does not exists and is now added in a raw form, not using
translation for now.

Usually translation **must** be done directly and still
we make a examption here due to the fact that reworking
the complete `EXT:academic_persons_edit` templates is
already scheduled and in the making including finishing
the translation.

This change is done in that way to have a working state
for development/testing purpose.