[BUGFIX] Use boolean to represent checkbox value
Because saving the form with an unchecked checkbox
results in an error this issue was investigated. It turns out
the `<f:form.checkbox>` ViewHelper adds an empty value
hidden input field which returns null when trying to set the
value for the model. As the set methods require an integer
value as parameter, trying to pass `null` results in this error.

Therefore a solution would have been allowing `null` to be
passed to the set methods and setting the property value to
`0` instead of `null`. After thinking of this solution it was
noticed that the `EXT:academic_persons_edit` also handles
checkbox values within a frontend based form, so this
approach is now used in the jobs as well.