[TASK] Add `targetFileNameWithoutExtension` for `FileUploadConverter`
The `FileUploadConverter` implementation provided by `EXT:academic_base`
has been introduced as a shared single file upload converter to reduce
duplicated code and implementation and `EXT:academic_jobs` has been used
for the initial implementation.

In preparation to replace the `EXT:academic_persons_edit` implementation
it has been revealed, that an additional configuration option is required
to allow providing a target filename pattern to be set instead of using
the uploaded filename completly, effectly ending in a replacement.

This change introduces `targetFileNameWithoutExtension` configuration as
`FileUploadConverter::CONFIGURATION_TARGET_FILE_NAME_WITHOUT_EXTENSION`
class constant and uses that pattern to enforce the target filename in
case the option is set.

Handles as task as this has been an oversight to be added directly and
due to the fact that the shared `EXT:academic_base` implementation is
new on the one hand and internal on the second hand anyway.