wip: !!! [TASK] Restructure persons edit templating
After the persons edit edit forms are now a kind of a step-by-step
form, proper templating is needed. This meant creating a
bootstrap compatible HTML including a useful filestructure. To
achieve this some controller actions do not have templates by
default as the actions are not used in the way the extension
wants the users to navigate through the edit forms. While
creating the frontend new icons needed to be introduced and
translation labels were adjusted not only in content, but also in
key to have a more meaningful and still easy naming convention.
Furthermore the name fields of the profile were set not only to
required, but also to disabled as the fields should not be
overwritten in the frontend. Also the breadcrumb navigation was
replaced by a back link, which should lead the user back to his
origin. While updating the templates, we also payed attention to
accessibility. To improve the back-link logic a few cancelUrls
needed to be added to the show views of the contracts and profile.