[TASK] ACE-33: Add `post-redirect-get(PRG)` for form persisting actions
It is good practice to use `POST`, `PUT`, `PATCH` and `DELETE`
http verbs (methods) for sending data to be persisted removing
them from server logs and mitigate simple get argument attacks
and link phishing.

That leads to a nifty side-effect that reloading the submitted
page will resend the data and perists it again or trying to
delete it, which seems to be reason to use `ForwardResponses()`
. This kind of response is a special implementation of TYPO3
Extbase MVC and only reroutes the request internally hidden
from clients, for example browsers.

The general technique to mitigate these kind of issues is called
`post-redirect-get` or in short `PRG`, which describes basically
that a POST request returns a redirect response to the same uri
using GET and discarding submitted data along with the original
request [1].

This change implements PRG now for `EXT:academic_persons_edit`
actions dealing with data persistance or manipulation to send
a redirect response instead of forwarding internally.

We Use `303 - see other` status code which is considerable the
semantically correct status code to tell the browser / client
to redirect to another uri and discard the POST data and not
sending it along with the redirect), which is what we want for
a `post-redirect-get (PRG)` implementation.

The same status code is also used for other redirect responses
within these controllers instead of `302` to follow semantic
correcet status code usage.

[1] https://en.ryte.com/wiki/Post-Redirect-Get/