[WIP][TASK] ACE-33: Enhance form data transformation and validation
-------------------------------------------------------------------
@todo's:

* Implement handling within factory classes to skip mapping
  data from form DTO to domain model in case property has
  not been sent within the current request.

* Investigate and implement a way to make "ignoring" not
  set property and still set them with the value set in
  the `AbstractFormData` as override.

-------------------------------------------------------------------

`EXT:academic_persons_edit` comes with a customized form and
validation handling, based on a the yaml-based configuration
system used to define validations for BE (FormEngine) and
in the frontend using extbase form handling provided using
form data transfer objects (AbstractFormData) combined with
transformation service (Factory - bad naming).

The combined concept lacked a couple of things, some of them
has already been fixed or correspondingly adjusted, leaving
proper checking during the data transformation that property
has been sent in persisting request or not. This change aims
to implement this missing piece, which comes additional with
the hurdle that property **may** not be sent in request but
still needs to be set, for example if PSR-14 events are put
in place before processing the transformation and fill-up
missing data based on other sources - this use-case needs
to be directly be considered !!!

* Added `AbstractFormData->setRequest()` method along with
  property `?ServerRequestInterface $request = null`, both
  required as preparation to modifiy DTO to domain model
  transformation (Factory) to skip applying values **not**
  transferred with the post request.

* Added `AbstractFormData->setArgumentName()` along with
  property `?string $argumentName = null`, both required
  as preparation to modify DTO to dmomain model mapping
  (Factory) so skip appliying values **not** transferred
  with the post request.

* Added `AbstractFormDataConverter` to allow setting the
  request to the `AbstractFormData` instance for mapped
  argument, as long as `request` is set with a valid
  `ServerRequestInterace` instance within corresponding
  TypeConverterOptions for the argument. Additionally,
  the `argumentName` is set if configured in TypeConverter
  config.

* Modify `AbstractActionController` calling the added
  `setCurrentRequestForAbstractFormDataBasedArguments()`
  method in `initializeActions()` to set the current
  request as TypeConverterOptions for the all arguments
  based on `AbstractFormData`. additionally the argument
  name is also set within the TypeConverterOptions, also
  required enhanced form data transformation handling.
  Sadly, there is no event or a more configuratble way
  to do that in TYPO3.