wip: [BUGFIX] Respect hidden profiles/contract when showing page related contacts
As the contacts4pages had been able to show the contract
related information of a hidden profile, they now are checked if
either the contract or the profile are not available to prevent
this issue.

Another approach would've been adjusting the findByPid
function of the ContactsRepository to not return relations of
hidden contracts or profiles in the first place.