[TASK] Ensure that select items are resolved correctly in JobController
`EXT:academic_jobs` provides two itemProcFuncs used to declare the
select items for `type` and `employe_type` available for backend
(FormEngine) also providing additional public methods for the items
directly. The latter method is used within the extbase controller
for the frontend plugin and the reason why the itemProcFunc classes
directly translated the item labels required for the frontend but
not for backend.

The controller used the items method directly and therefore has
the issue that this does not respect manually configured items
using TCA/TCAOverride or repecting replaced itemProcFunc config,
for example done in project to extend/modify the items.

This change streamlines the implementation now:

* `EmploymentTypeItems` and `TypeItems` are modified making the
   default items methods private and removing direct translating
   item labels.

* Introduce `GetSelectItemsForTcaManagedTableFieldMethodTrait`
  in `EXT:academic_base` to provide a reusable method to be
  used in controllers return the tca select items. This trait
  could later also be used in extension to retrieve select
  items the same field, for example if additional select fields
  has been added.

* `JobController` removes the type classes (DI), adding localization
  utility to be injected and replace the field item retrieven for
  the view using the `EXT:academic_base` method trait.