[TASK] Add `EXT:academic_base` internal `ArrayObjectMapper` service
> [!IMPORTANT]
> The introduced service and related classes are purely internal
> for `EXT:academic_base` and not part of `academic-*` extension
> internal API and even more not as public API to be used by 3rd
> party extensions, which should implement their own package.
>
> Due to the completly internal declaration this service is not
> documented (yet) in any form and also considerable experimental
> with ongoing improvements and adjustments when they rise up.

This change introduces the new internal service `ArrayObjectMapper`,
based on the `symfony/serializer` component as preparation to provide
a `academic/*` extension internal customizable settings system mapped
to extension setting classes.

The required serializer configuration is provided by a more or less
static composed `SerializerFactory` used to configure a internal
service configuration injected into `ArrayObjectMapper` service,
specialized for the internal and tested use-case.

Based on the required dependency and the fact that this component
is not provided by TYPO3, neighter direct or indirect in composer
and classic mode installations, the package is added as dependency
and also added to the extra configuration to have it in place so it
can be bundled for classic mode installation uploads to the TYPO3
extension repository (TER) using tailer.

Background for the decision to implement a custom service stems
mainly from the fact that the TYPO3 Extbase `PropertyMapper` is
not really use-full for this scenario and requires to provide a
lot of custom `TypeConverters` for custom classes and having huge
problems with array-of-objects provided as list-arrays without a
working implementation for it. Not requireing dependency injection
and ruleing out X'Classing or any kind of replacement as design
decision was the birth of the custom mapper service idea and now
implementation.

To visualize the goal of the introduced internal service, the
following (simplified) example object structure is assumed:

```php
<?php

namespace MyVendor\MyExtension\Settings;

final class Features
{
  public function __construct(
    public readonly bool $featureOne,
    public readonly bool $featureTwo,
    public readonly bool $featureThree,
  ) {}
}

final class Field
{
  public function __construct(
    public readonly string $identifier,
    public readonly bool $required,
    public readonly bool $disabled,
  ) {}
}

final class CustomExtensionSettings
{
  /**
   * @param Field[] $fields
   **/
  public function __construct(
    public readonly Features $features,
    public readonly array $fields,
  ) {}
}
```

Further we assume to have for example following array:

```php
<?php
$settingsArray = [
  'features' => [
    'featureOne' => true,
    'featureTwo' => false,
    'featureThree' => false,
  ],
  'fields' => [
    [
      'identifier' => 'field_one',
      'required' => false,
      'disabled' => false,
    ],
    [
      'identifier' => 'field_two',
      'required' => true,
      'disabled' => false,
    ],
  ],
];
```

The new service can now be used to map the assumped array to
the assumped `CustomExtensionSettings` object creating the
nested object structure using:

```php
<?php

// Prefere dependency injection, used only for simplicity
$arrayObjectMapper = GeneralUtility::makeInstance(
  ArrayObjectMapper::class,
);
$customExtensionSettings = $arrayObjectMapper->map(
  $data,
  CustomExtensionSettings::class,
);
```

and everything goes well we have the full object structure
of `CustomExtensionSettings` with the applied data.

Be aware that the above example is only a really basic one
and because we are using the `symfony/serializer` with a
generic configuration set we already support other class
designes as well, which can be looked up in the included
functional tests and fixture implementations. To peak into
some of them:

* Setting data to private/protected properties are possible
  even without constructor arguments.

* Constructor Parameter Promotion (PCC) is supported like
  used in the above example, with `public`, `protected`
  and `private` visibility.

* Setter method are used if available an no constructure
  argument/constructor exists for that property.

* Native types and phpdoc-block annotation are respected.

* `symfony/serializer` and `symfony/property-access` php
  attributes can be used.

The factory and implemented mapping feature may be adjusted
and/or extended over the time and is the reason why this is
considered highly experimental and internal nailed down to
`EXT:academic_base` for upcoming features and not designed
to be used standalone. Can chane at anytime without being
considered breaking.

To give a rough overview, the service provided by this change
**should** be the fundament for an automatic `YAML` to object
service for our `academic-*` extensions based on simplified
configuration, allowing multiple settings-to-object for each
extension powerd directly with a `root-package` AND override
by other extensions backed by a centralized DI factory taking
care of instantiation, caching and injection as replacement
for the used settings system of `EXT:academinc_persons` and
required for other extensions. Having a generic way instead
of reimplementing the same over and over again will rock, so
stay tuned for it.

> [!NOTE]
> The bundling process is not included and scheduled to
> be implemented based on the extra configuration as an
> automatic process using either a monorepository internal
> helper package or a dedicated package/tool, which is
> scheduled to been developed but concrete implementation
> not figured out - except the definition of the package
> configuration used to use composer for bundle package
> installation.

The implementation has been build on modern and recommended
design principles regarding class designs, service designs
and concern of separation, following modern best-practice.

Used command(s):

```shell
Build/Scripts/runTests.sh -s composer -- config \
  -d packages/fgtclb/academic-base \
  --json "extra"."sbuerk/release"."tailor" \
    '{"require":{"symfony/serializer":"^6.4.26"}}' \
&& Build/Scripts/runTests.sh -s composer -- require \
  -d packages/fgtclb/academic-base --no-update \
    'symfony/serializer':'^6.4.26 || ^7.3.4'
```

[1] https://symfony.com/doc/6.4/serializer.html