[DOCS] Add changelog templates
In preparation of structured changelog rendering we need templates for
ReST files. As we will go the same way the TYPO3 core went before we'll
use 4 types of changelog messages:

* Breaking changes
* Deprecations
* Features
* Important Notes

These messages use different sections for describing the effective
changes.

For an easy start, this commit adds templates for the messages.
They are located in the `Build` folder because they made are for
internal use and won't get shipped nor released.

See:

* https://github.com/TYPO3/typo3/blob/main/typo3/sysext/core/Documentation/Changelog/Howto.rst