[DOCS] Add changelog structure for Category Types extension
As this extension follows the TYPO3 core way of documenting changes a
basic files/folder structure is needed. This is added in this change.

We'll put the special types of change documentations in a folder named
by the minor version they are/were introduced. Additionally there are
three more general files which are including the single change
documents:

- `Documentation/Changelog/Changelog-${major}.rst`
- `Documentation/Changelog/Changelog-${major}-combined.rst`
- `Documentation/Changelog/${major}.${minor}/Index.rst`

For each new major version change the corresponding files
`Documentation/Changelog/Changelog-${major}.rst` and
`Documentation/Changelog/Changelog-${major}-combined.rst` must be
created and linked to in `Documentation/Index.rst`.

For each new minor version change a corresponding folder must be
created and linked to in
`Documentation/Changelog/Changelog-${major}.rst`.