[FEATURE] Added functionality to display only “Show Public” profile.
This gives profile a setting option to determine whether their records may be used publicly.
To this end, a change was made to the plugin controller.

A DomainGenerator was introduced to better manage the plugin settings than just in the controller.
The `ProfileController` is used for different plugins that have different FlexForm configurations.
By inserting the DomainGenerator as a layer in between, a distinction can be made between configurations from `ContractSelector` and `ProfileList`.

A new field called `allowed_show_public` was introduced for the profiles table.
In the current implementation, however, the public visibility of contracts is determined by the profiles in the `ContractSelector` plugin.
Profile and Contract records may only be displayed publicly in the area if they are allowed via this record.