[BUGFIX] ACE-116: Use correct PageTSConfig includes for `EXT:academic_programs`
During reviewing, streamlining and testing PR #242 it
was revealed that the `EXT:academic_programs` plugins
are not listed within the `academic` group or listed
at all in the `New Content Element` wizard and relates
to a missing PageTsConfig import in TYPO3 v12 and v13
`sys_template` setup.

This change adds the correct import and removes one
import not matching any files and displayed within
the `PageTsConfig` module as syntax error.