[TASK] ACE-118: Add PSR-14 to modify `StateBuildContext`
This change adds the following internal PSR-14 event
`ModifyProfileCreateEnvironmentStateBuildContextForFrontendUserEvent`
dispatched to allow project to modify the environment
build context on per frontend user record basis, or
if needed to prevent bootstrapping a custom environment.

The event is marked as `@internal` due to it's experimantal
state like the whole Environment State management and
build feature, which may be extracted into a dedicated
extension at some point.

Main reason for adding the internal event is having a
quick and possible way at hand in project if default
environment build context setup does not work out.